multi updates

+++ app/Admin/Controllers/InstitutesController.php
@@ -18,7 +18,8 @@ use Encore\Admin\Admin;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Storage;
 use App\Models\Report;
-use App\Models\ReportFiles;
+use App\Models\ReportsFiles;
+use App\Admin\Extensions\Tools\PdfButton;
 class InstitutesController extends Controller
 {
 
@@ -113,13 +114,15 @@ class InstitutesController extends Controller
             'name' => 'required|string|max:255',
             'phone' => 'required|string|max:18',
             'email' => 'required|string|email|max:255',
-            'password'=>'required|string|min:6|confirmed',
-            'username'=>'required|string|max:25',
+            'password'=>'nullable||string|min:6|confirmed',
+            'username'=>'string|max:25',
             'logo' => 'nullable|image|max:1000|dimensions:min_width=150,min_height=150|mimes:jpeg,png,gif'
         ]);
         $data=$request->all();
         if(!empty($data['password'])){
             $data['password'] = bcrypt($data['password']);
+        }else{
+            unset($data['password']);
         }
         /**upload and save image***/
         $uploadedFile = $request->file('logo');
@@ -222,7 +225,10 @@ SCRIPT;
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
+    {   $reportexits=Report::where('institute_id','=',$id)->count();
+        if($reportexits){
+            return response()->JSON(['status' => false, 'message' => 'Selected institute having report records, so you cant delete this institute.']);
+        }
         if(Institute::destroy($id)){
             $res = ['status' => true, 'message' => 'Institute has been removed.'];
         }else{
@@ -244,7 +250,10 @@ SCRIPT;
         $show = new Show(Institute::findOrFail($id));
         $show->panel()
         ->style('danger')
-        ->title('Account Details');
+        ->title('Account Details')
+        ->tools(function ($tools) {
+            $tools->prepend(new PdfButton());
+        });
         $show->name(trans('Name'));
         
         $show->username(trans('username'));
@@ -268,10 +277,14 @@ SCRIPT;
     }
     protected function institutedetail($id)
     {
-
-        $show = new Show(Institute::findOrFail($id));
+        $institute=Institute::findOrFail($id);
+        $show = new Show($institute);
         $show->panel()
-        ->title('Institute Details');
+        ->title($institute->name)
+        ->tools(function ($tools) {
+            $tools->prepend(new PdfButton());
+        });
+        
         $show->logo()->image();
         $show->client_male(trans('Client Male'));
         $show->client_female(trans('Client Female'));
@@ -446,13 +459,18 @@ protected function reportgrid($id)
 {
 
     $grid = new Grid(new Report());
+    $grid->tools(function ($tools) {
+        $tools->disableRefreshButton();
+        //$tools->prepend(new PdfButton());
+    });
     $grid->model()->where('institute_id','=',$id)->orderBy('id', 'DESC');
     $grid->paginate(20);
-    $grid->disableExport();
-    $grid->disableActions();
+    //$grid->disableExport();
+    //$grid->disableActions();
+    //$grid->disableRefreshButton();
     $grid->disableFilter();
     $grid->disableRowSelector();
-    $grid->disableCreateButton();
+    //$grid->disableCreateButton();
     $grid->id('ID');
     $grid->report_category(trans('Report Category'));
     $grid->column('Period')->display(function () {
@@ -466,7 +484,7 @@ protected function reportgrid($id)
             return '';
         }
     });
-    $grid->report_year(trans('Year'));
+    //$grid->report_year(trans('Year'));
     $grid->total_capital(trans('Total Capital'));
     $grid->total_assest(trans('Total Assest'));
     $grid->total_liability(trans('Total Liability'));
@@ -476,9 +494,307 @@ protected function reportgrid($id)
     $grid->return_average_assets(trans('Return Ave Assets'));
     $grid->return_equity(trans('Return Equity'));
 
-    $grid->created_at(trans('Created'))->display(function($date){
+    /*$grid->created_at(trans('Created'))->display(function($date){
         return CommonMethod::formatDate($date);
-    });
+    });*/
     return $grid;
 }
+
+/***export PDF ***/
+function export($id){
+    $institute=Institute::findOrFail($id);
+    if(Storage::disk('user_uploads')->exists($institute->logo)){
+        $logo = '/uploads/user/' . $institute->logo ;
+    }else{
+        $logo = '';
+    }
+    $pocessedHtml=view('admin.institutes.export_pdf',['institute'=>$institute,'logo'=>$logo])->render();
+    include(base_path('vendor/xtcpdf.php'));
+    $tcpdf = new \XTCPDF();
+    $textfont = 'freesans'; // looks better, finer, and more condensed than 'dejavusans'
+
+    $tcpdf->SetAuthor("COBAT");
+    //$tcpdf->SetAutoPageBreak( false );
+    $tcpdf->SetAutoPageBreak(true, 40);
+
+    $tcpdf->setHeaderFont(array($textfont,'',30));
+    $tcpdf->xheadercolor = array(255,255,255);
+    $tcpdf->xheadertext = '';
+    $tcpdf->xfootertext = 'COBAT IMS';
+    $tcpdf->AddPage();
+    $tcpdf->SetTextColor(0, 0, 0);
+    $tcpdf->SetFont($textfont,'A',15);
+    $tcpdf->WriteHTML($pocessedHtml);
+    $filePath =$institute->username.'.pdf';
+    $tcpdf->Output($filePath, 'D');
+    //return response()->download($tcpdf->Output($filePath, 'I'), $filePath, ['Content-Type' => 'application/pdf']);
+    die;
+}
+
+/**institute report create**/
+public function createreport($id,Content $content)
+    { 
+        //return view('Clients.create');
+        return $content
+        ->header('Create Report')
+        ->description('Create...')
+        ->body(view('admin.institutes.reportadd',['url'=>'/admin/institutes/'.$id.'/createreportsave','id'=>$id]))
+        ->breadcrumb(
+            ['text' => 'Institute', 'url' => '/institutes'],
+            ['text' => Institute::find($id)->name,'url'=>'/institutes/'.$id],
+            ['text' => 'Create Report']
+        );
+        
+    }
+public function createreportsave($id,Request $request)
+    {   
+        $valid = request()->validate([
+            'report_category' => 'required',
+            'submission_period' => 'required',
+            'total_capital' => 'required|nullable|numeric',
+            'total_assest' => 'required|nullable|numeric',
+            'total_liability' => 'required|nullable|numeric',
+            'loan_advance' => 'required|nullable|numeric',
+            'customer_deposits' => 'required|nullable|numeric',
+            'profit_before_tax' => 'required|nullable|numeric',
+            'return_average_assets' => 'required|nullable|numeric',
+            'return_equity' => 'required|nullable|numeric',         
+            //'files' => 'mimes:jpeg,png,gif,pdf,doc,docx'
+        ]); 
+        $data=$request->all();
+        $submission_period = ($data['report_category']=="Monthly")?$data['submission_period']:NULL;
+        $submission_quater = ($data['report_category']=="Quaterly")?$data['submission_period']:NULL;        
+        $report_year = ($data['report_category']=="Audited")?$data['submission_period']:NULL;
+        
+        //check if already submitted
+        $existObj=Report::where('institute_id','=',$id)->where('report_category','=',$data['report_category']);
+        if($data['report_category']=="Monthly"){
+            $existObj->where('submission_period','=',$data['submission_period']);
+        }
+        if($data['report_category']=="Quaterly"){
+            $existObj->where('submission_quater','=',$data['submission_period']);
+        }
+        if($data['report_category']=="Audited"){
+            $existObj->where('report_year','=',$data['submission_period']);
+        }
+        if($existObj->exists()){
+            echo "<pre>";print_r($existObj->get());die;
+            admin_error('Error','Report already submitted for selected period');
+            return back()->with('Error','Report already submitted for selected period');    
+        }
+        $report = new Report;
+        $report->institute_id =$id;
+        $report->report_category = $data['report_category'];
+        $report->submission_period = $submission_period ;
+        $report->submission_quater = $submission_quater ;
+        $report->report_year = !empty($report_year)?$report_year:date('Y');
+        $report->total_capital = $data['total_capital'];
+        $report->total_assest  = $data['total_assest'];
+        $report->total_liability = $data['total_liability'];
+        $report->loan_advance = $data['loan_advance'];
+        $report->customer_deposits = $data['customer_deposits'];
+        $report->profit_before_tax = $data['profit_before_tax'];
+        $report->return_average_assets = $data['return_average_assets'];
+        $report->return_equity = $data['return_equity'];    
+
+        if($report->save()){                                
+
+            foreach ($request->files as $key => $attFiles) {
+                foreach ($attFiles as $fileData) {              
+                    if($fileData && $fileData->isValid()){
+                        $filename = time().'-'.$fileData->getClientOriginalName();
+                        $file = Storage::disk('user_docuploads')->putFileAs('',$fileData,$filename);
+
+                        $reportFiles = new ReportsFiles;
+                        $reportFiles->report_id =  $report->id;              
+                        $reportFiles->filename = $file;
+                        $reportFiles->save();       
+                    }                   
+                }
+            }
+            /**upload and save image end here***/
+            admin_success('Success','Report has been successfully added!');
+            return redirect('/admin/institutes/'.$id);
+        }else{
+            admin_error('Error','Something went wrong! Please Try Again.');
+            return back();
+        }
+    }
+
+/**edit report for selected institute***/
+public function editreport($id,$reportid,Content $content)
+    { 
+        //return view('Clients.create');
+        $report=Report::findOrFail($reportid);
+        return $content
+        ->header('Edit Report')
+        ->description('Edit...')
+        ->body(view('admin.institutes.reportedit',['report'=>$report,'url'=>'/admin/institutes/'.$id.'/'.$reportid.'/editreportsave','id'=>$id]))
+        ->breadcrumb(
+            ['text' => 'Institute', 'url' => '/institutes'],
+            ['text' => Institute::find($id)->name,'url'=>'/institutes/'.$id],
+            ['text' => 'Edit Report']
+        );
+        
+    }
+ /**edit report save**/
+ public function editreportsave($id,$reportid,Request $request){
+
+        $institute = request()->validate([
+            'report_category' => 'required',
+            'submission_period' => 'required',
+            'total_capital' => 'required|nullable|numeric',
+            'total_assest' => 'required|nullable|numeric',
+            'total_liability' => 'required|nullable|numeric',
+            'loan_advance' => 'required|nullable|numeric',
+            'customer_deposits' => 'required|nullable|numeric',
+            'profit_before_tax' => 'required|nullable|numeric',
+            'return_average_assets' => 'required|nullable|numeric',
+            'return_equity' => 'required|nullable|numeric',
+        ]);
+        $data=$request->all();
+        $submission_period = ($data['report_category']=="Monthly")?$data['submission_period']:NULL;
+        $submission_quater = ($data['report_category']=="Quaterly")?$data['submission_period']:NULL;        
+        $report_year = ($data['report_category']=="Audited")?$data['submission_period']:NULL;
+        
+        $report =  Report::find($data['id']);
+        $report->report_category = $data['report_category'];
+        $report->submission_period = $submission_period ;
+        $report->submission_quater = $submission_quater ;
+        $report->report_year = !empty($report_year)?$report_year:date('Y');
+        $report->total_capital = $data['total_capital'];
+        $report->total_assest  = $data['total_assest'];
+        $report->total_liability = $data['total_liability'];
+        $report->loan_advance = $data['loan_advance'];
+        $report->customer_deposits = $data['customer_deposits'];
+        $report->profit_before_tax = $data['profit_before_tax'];
+        $report->return_average_assets = $data['return_average_assets'];
+        $report->return_equity = $data['return_equity'];    
+
+        if($report->save()){                                
+
+            foreach ($request->files as $key => $attFiles) {
+                foreach ($attFiles as $fileData) {              
+                    if($fileData && $fileData->isValid()){
+                        $filename = time().'-'.$fileData->getClientOriginalName();
+                        $file = Storage::disk('user_docuploads')->putFileAs('',$fileData,$filename);
+
+                        $reportFiles = new ReportsFiles;
+                        $reportFiles->report_id =  $report->id;              
+                        $reportFiles->filename = $file;
+                        $reportFiles->save();       
+                    }                   
+                }
+            }
+            //notifyt to Admin
+            //$this->notifytoadmin($profile, $data);
+
+            admin_success('Success','Report has been successfully updated!');
+            return redirect('/admin/institutes/'.$id);
+        }else{
+            admin_error('Error','Something went wrong! Please Try Again.');
+            return back();
+        }
+    }  
+    /***delete report ****/
+    public function deletereport($id,$reportid)
+    {   
+        if(Report::destroy($reportid)){
+            ReportsFiles::where('report_id', $reportid)->delete();
+            $res = ['status' => true, 'message' => 'Report has been removed.'];
+        }else{
+            $res = ['status' => false, 'message' => 'Something went wrong!'];
+        }
+        return response()->JSON($res);
+    }  
+
+    public function viewreport($id,$reportid ,Content $content)
+    {    
+        return $content
+        ->header('Report Details')
+        ->description('Edit...')
+        ->body($this->reportdetail($reportid)->render())
+        ->breadcrumb(
+            ['text' => 'Institute', 'url' => '/institutes'],
+            ['text' => Institute::find($id)->name,'url'=>'/institutes/'.$id],
+            ['text' => 'Report']
+        );
+        
+        
+    }
+    protected function reportdetail($id)
+    {
+        $script = <<<SCRIPT
+
+        window.onload = function () {
+            jQuery('.trigger').click(function(){
+                var filename=jQuery(this).attr('rel-file');
+                jQuery("#myModal .modal-body").find('iframe').attr('src','/uploads/user/doc/'+filename);
+                jQuery("#myModal").modal('show');
+            }); 
+        }
+SCRIPT;
+        Admin::script($script); 
+        $show = new Show(Report::findOrFail($id));
+        $show->panel()
+        ->style('danger')
+        ->title('Report Details')
+        ->tools(function ($tools) {
+            $tools->disableDelete();
+        });
+        $show->institute_id(trans('Institue'))->as(function($id){
+            return Institute::find($id)->name;
+        });
+        $show->report_category(trans('Report Category'));
+        $show->contents('Period')->as(function () {
+            if($this->report_category=='Monthly'){
+                return CommonMethod::getMonthName($this->submission_period);
+            }
+            if($this->report_category=='Quaterly'){
+                return CommonMethod::getQuarterName($this->submission_quater);
+            }
+            if($this->report_category=='Audited'){
+                return '';
+            }
+        });
+        $show->report_year(trans('Year'));
+        $show->total_capital(trans('Total Capital'));
+        $show->total_assest(trans('Total Assest'));
+        $show->total_liability(trans('Total Liability'));
+        $show->loan_advance(trans('Loan Advance'));
+        $show->customer_deposits(trans('Customer Deposits'));
+        $show->profit_before_tax(trans('Profit Exc Tax'));
+        $show->return_average_assets(trans('Return Ave Assets'));
+        $show->return_equity(trans('Return Equity'));
+        
+        $show->created_at(trans('Created'))->as(function($date){
+            return CommonMethod::formatDate($date);
+        });
+
+        $show->contents(trans('Financial Report'))->setEscape(false)->as(function($date){
+            $return= '<div id="myModal" class="modal fade" role="dialog">
+                  <div class="modal-dialog">
+                    <div class="modal-content">
+                      <div class="modal-header">
+                        <button type="button" class="close" data-dismiss="modal">&times;</button>
+                      </div>
+                      <div class="modal-body" style="min-height:600px;">
+                        <iframe src"" width="100%" height="100%" style="min-height:550px;"></iframe>
+                      </div>
+                      
+                    </div>
+
+                  </div>
+                </div>';
+            $reportfiles=ReportsFiles::where('report_id','=',$this->id)->get();
+            $return .='<ul>';
+            foreach($reportfiles as $reportfile):
+                $return .='<li class="trigger" rel-file="'.$reportfile->filename.'">'.$reportfile->filename.'</li>';
+            endforeach;
+            $return .='</ul>';
+            return $return;
+        });
+
+        return $show;
+    } 
 }
